Refactoring graphs and charts.
