+++ gui-flask/static/js-qwak/js-dynamic/dynamic-tools.js
@@ -108,7 +108,6 @@ document.getElementById('customGraphDisplay').addEventListener('click', evt => {
 
 })
 
-
 export let cyDynamic = cytoscape({
     container: document.getElementById("cyDynamic"), // container to render in
     boxSelectionEnabled: false,
@@ -241,187 +240,6 @@ export let customCyDynamic = cytoscape({
         ]
     }
 });
-
-export let dynamicChartData = {
-    type: "line",
-    data: {
-        labels: [...Array(100).keys()],
-        datasets: [
-            {
-                label: "Probability",
-                data: [],
-                fill: false,
-                borderColor: "rgb(14,84,246)",
-                pointRadius: 0,
-            },
-        ],
-    },
-    options: {
-        scales: {
-            x: {
-                grid: {
-                    display: false
-                }
-            },
-            y:{
-                suggestedMax:0.1,
-                suggestedMin:0,
-            },
-        },
-    },
-};
-
-export let dynamicMeanChartData = {
-    type: "line",
-    data: {
-        labels: [...Array(100).keys()],
-        datasets: [
-            {
-                label: "Node",
-                data: [],
-                fill: false,
-                borderColor: "rgb(75, 192, 192)",
-                pointRadius: 0,
-            },
-        ],
-    },
-    options: {
-        scales: {
-            x: {
-                grid: {
-                    display: false
-                }
-            },
-        },
-    },
-};
-
-export let dynamicStDevChartData = {
-    type: "line",
-    data: {
-        labels: [...Array(100).keys()],
-        datasets: [
-            {
-                label: "Node",
-                data: [],
-                fill: false,
-                borderColor: "rgb(75, 192, 192)",
-                pointRadius: 0,
-            },
-        ],
-    },
-    options: {
-        scales: {
-            x: {
-                grid: {
-                    display: false
-                }
-            },
-        },
-    },
-};
-
-export let dynamicInvPartRatioChartData = {
-    type: "line",
-    data: {
-        labels: [...Array(100).keys()],
-        datasets: [
-            {
-                label: "Node",
-                data: [],
-                fill: false,
-                borderColor: "rgb(75, 192, 192)",
-                pointRadius: 0,
-            },
-        ],
-    },
-    options: {
-        scales: {
-            x: {
-                grid: {
-                    display: false
-                }
-            },
-        },
-    },
-};
-
-export let dynamicSurvivalProbChartData = {
-    type: "line",
-    data: {
-        labels: [...Array(100).keys()],
-        datasets: [
-            {
-                label: "Node",
-                data: [],
-                fill: false,
-                borderColor: "rgb(75, 192, 192)",
-                pointRadius: 0,
-            },
-        ],
-    },
-    options: {
-        scales: {
-            x: {
-                grid: {
-                    display: false
-                }
-            },
-        },
-    },
-};
-
-export let staticChartData = {
-    type: "line",
-    data: {
-        labels: [...Array(defaultDist.length).keys()],
-        datasets: [
-            {
-                label: "Probability",
-                data: defaultDist,
-                borderWidth: 1,
-                fill: false,
-                borderColor: "rgb(21, 52, 228)",
-                tension: 0.1,
-                pointRadius: 0,
-            },
-        ],
-    },
-    options: {
-        responsive: true,
-        maintainAspectRatio: false,
-        scales: {
-            y: {
-                beginAtZero: true,
-            },
-        },
-
-    },
-};
-
-document.getElementById('defaultDynStat').addEventListener('click', evt => {
-    openTab(evt, 'Mean', "stattabcontent", "stattablinks");
-});
-
-document.getElementById('stDevDynStat').addEventListener('click', evt => {
-    openTab(evt, 'StDev', "stattabcontent", "stattablinks");
-});
-
-document.getElementById('invPartRatioDynStat').addEventListener('click', evt => {
-    openTab(evt, 'InvPartRatio', "stattabcontent", "stattablinks");
-});
-
-document.getElementById('survProbDynStat').addEventListener('click', evt => {
-    openTab(evt, 'SurvivalProb', "stattabcontent", "stattablinks");
-});
-
-document.getElementById('invPartRatioDynStat').click()
-
-document.getElementById('invPartRatioDynStat').click()
-
-document.getElementById('stDevDynStat').click()
-
-document.getElementById('defaultDynStat').click()
 
 document.getElementById('defaultOpen').click()
