+++ gui-flask/templates/dynamicQW.html
@@ -37,7 +37,7 @@
 <div class="parent-wrapper">
     <!-- Header -->
     <header id="header">
-        <a href="index.html" class="title">QWAK</a>
+        <a href="https://github.com/JaimePSantos/qwak" class="title">QWAK</a>
         <nav>
             <ul>
                 <li><a href="{{url_for('home')}}">Home</a></li>
@@ -118,14 +118,14 @@
                                 </ul>
                                 <ul class='input-fields'>
                                     <li class='input-combo2'>
-                                        <div id="placeholderDiv" class="myText probDistInput">NA</div>
+                                        <div id="placeholderDiv" class="myText probDistInput">[NA]</div>
                                         <input type="text" name="placeholder" id="placeholderInp"
                                                        class="myProbDistInputBox"></input>
                                     </li>
                                     <li class='input-combo'>
                                         <input type="text" step="0.01" min=0 name="placeholder" id="placeholderInp2"
                                                        class="input-box"></input>
-                                        <div id="placeholderDiv2" class="myText probDistInput">NA</div>
+                                        <div id="placeholderDiv2" class="myText probDistInput">[NA]</div>
                                     </li>
                                 </ul>
                             </div>
