+++ gui-flask/templates/staticQW.html
@@ -28,7 +28,7 @@
             crossorigin="anonymous" referrerpolicy="no-referrer"></script>
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
     <script type='module' src="{{ url_for('static',filename='js-qwak/static-script.js') }}"></script>
-    <script type='module' src="{{ url_for('static',filename='js-qwak/js-static/static-graph.js') }}"></script>
+    <script type='module' src="{{ url_for('static',filename='js-qwak/graphs.js') }}"></script>
     <script type='module' src="{{ url_for('static',filename='js-qwak/js-static/static-probDist.js') }}"></script>
     <script type='module' src="{{ url_for('static',filename='js-qwak/js-static/static-tools.js') }}"></script>
 
@@ -39,7 +39,7 @@
 <div class="parent-wrapper">
     <!-- Header -->
     <header id="header">
-        <a href="index.html" class="title">QWAK</a>
+        <a href="https://github.com/JaimePSantos/qwak" class="title">QWAK</a>
         <nav>
             <ul>
                 <li><a href="{{url_for('home')}}">Home</a></li>
